+++ js/modules/posts.js
@@ -1,0 +1,20 @@
+import {addDescription, getRandomInteger} from './util.js';
+import {POST_MIN_LIKES, POST_MAX_LIKES, COMMENT_MAX_COUNT, COMMENT_MIN_COUNT} from './variables.js';
+import {addComment} from './comments.js';
+
+
+const addPost = (_, index) => {
+  const id = index + 1;
+  return {
+    id,
+    url: `photos/${id}.jpg`,
+    description: addDescription(),
+    likes: getRandomInteger(POST_MIN_LIKES, POST_MAX_LIKES),
+    comments: Array.from({ length: getRandomInteger(COMMENT_MIN_COUNT, COMMENT_MAX_COUNT) }, addComment),
+  };
+};
+
+const posts = () => Array.from({ length: 25 }, addPost);
+
+
+export {addPost, posts};
