+++ js/modules/variables.js
@@ -1,0 +1,61 @@
+const COMMENT_AUTHOR_NAMES = [
+  'Илья',
+  'Михаил',
+  'Елена',
+  'Софья',
+  'Татьяна',
+  'Кекс',
+  'Рудольф',
+  'Александр',
+  'Саня',
+  'Илюха',
+  'Денис',
+  'Даниил',
+  'Иван',
+  'Ольга',
+  'Алексей',
+  'Инокентий',
+];
+
+const TEXT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const TEXT_DESCRIPTION = [
+  'Горы Кавказа',
+  'Фотография меня с котом',
+  'Подскользнулся))',
+  'Красивые водопады',
+  'Сердце замерло',
+  'Красота',
+  'На чилле, на расслпбоне',
+  'Работаем братья',
+  'Санечка снимает',
+  'Ой',
+  'Милый котик',
+  'Гроза района',
+  'Папина дочка',
+  'Дома',
+  'С собакой на прогулке',
+  ')))))',
+  'Вид из окна',
+  'Как-то так',
+  '...',
+  'На концерте'
+];
+
+const USER_AVATAR_MIN_ID = 1;
+const USER_AVATAR_MAX_ID = 6;
+const POST_MIN_LIKES = 15;
+const POST_MAX_LIKES = 200;
+const COMMENT_MIN_COUNT = 1;
+const COMMENT_MAX_COUNT = 5;
+const COMMENT_ID_START = 10;
+const COMMENT_ID_STEP = 5;
+
+export {COMMENT_AUTHOR_NAMES, TEXT_MESSAGES, TEXT_DESCRIPTION, USER_AVATAR_MIN_ID, USER_AVATAR_MAX_ID, POST_MIN_LIKES, POST_MAX_LIKES, COMMENT_MIN_COUNT, COMMENT_MAX_COUNT, COMMENT_ID_START, COMMENT_ID_STEP};
